[libc++] Fix std::type_info comparison
As noted by Richard Smith in https://reviews.llvm.org/D97802#2675067, the
current implementation of std::type_info less-than comparison was not an
ordering relation. This commit fixes that.

It's technically a behavior change, however anyone relying on the ordering
as given by the previous implementation was relying on something extremely
brittle, so I think it's safe to fix this.

Differential Revision: https://reviews.llvm.org/D100134