[WIP][libc++] Implement ABI list tests in the test suite
This commit moves the ad-hoc testing previously implemented in CMake
into the Lit test suite, integrating it seamlessly with the rest of the
test suite.

This is the first Lit test that specifically checks properties of the
shared library we build, and as such, it provides the framework to write
additional tests of the same kind. That is something that we've previously
struggled to do in libc++'s test suite, and for which I know at least Apple
has needed a custom downstream solution to implement. This should provide
a place for these tests to live for any vendor.