[lit] Use getFilePath() to access the physical file backing a Lit test
lit.Test objects take a file_path argument in their constructor. That
seems to be used to provide a file that differs from the actual source
path of the test. However, the getFilePath() method wasn't used in the
places where Lit is expecting to read an actual file on disk, leading to
issues when creating lit.Tests with a custom file_path.

The custom file_path argument doesn't seem to be used within Lit,
however I found a use case that requires it: it is possible to generate
tests on the fly during test suite execution, and doing so requires using
a physical file that differs from the test's source path (otherwise we'd
have to generate the files in the source tree, which is bad).