[libc++] Unify the benchmarks with the test suite
Instead of building the benchmarks separately via CMake and running them
separately from the test suite, this patch merges the benchmarks into
the test suite and handles both uniformly.

As a result:
- It is now possible to run individual benchmarks like we run tests,
  which is a huge quality-of-life improvement.

- The benchmarks will be run under exactly the same configuration as
  the rest of the tests. This does mean that one has to be careful to
  enable the desired optimization flags when running benchmarks, which
  is easy with e.g. `lit <...> --param optimization=speed`.

- Benchmarks can use the same annotations as the rest of the test
  suite, such as `//UNSUPPORTED` & friends.

- It is not possible anymore (for now) to disable the benchmarks while
  enabling the test suite. This may be added in the future if there is
  a use case for it.

When running the tests, it is usually not necessary for correctness to
run the full benchmarks. Running the full benchmarks inside `check-cxx`
would only be very time consuming for little benefit. As a result, we
only run each benchmark for a very short amount of time when triggered
via `check-cxx`.

TODO:
- Refactor the implementation of the format for `.bench.cpp`. We'll need
  a way to pass through the path to google benchmark, and we'll also need
  a way to customize the argument we pass to the benchmark at runtime for
  testing vs actually benchmarking.