[libc++] Enable more features when localization is disabled
When localization is disabled, implement the localization layer by
assuming the C locale instead of bailing out. This allows several parts
of the library like regex and streams to just work out of the box when
localization is disabled.

It also greatly simplify the codebase, which had accumulated a very
large number of conditionals to exclude parts that used to "require"
localization, when in reality there was nothing specific to localization
in that code.

This makes the "no-localization" configuration closer to a
"only-the-C-locale-is-supported" configuration, which is more
generally useful.

We also now omit to include `<locale.h>` from `<clocale>` when
localization is disabled. This ensures that the library does not
attempt to rely on localization support from the C library when
it has been configured not to.