[libc++] Remove the legacy localization base API
Now that all officially supported platforms have been moved to the
new API, we can remove the legacy API. Note that this also removes
support for OpenBSD and Musl from the localization base API, however
these platforms are not officially supported and we have no CI coverage.

If someone on these platforms is broken by this patch, please reach
out and we can provide assistance to define the locale base API on
these platforms (and add CI!).